Use None to signalize the end of the appended data
Since we need to use a channel to allow to append data into the
source nodes, we need to inform the nodes about the data being processed
for the chunk of data provided. Once the chunk of data is sent to
the outputs of the flow tree, these ones would generate the visual
marks for that data chunk.