Implement multiple sources
To allow the flow graph to propagate values across different receiving
nodes, the source node needs to be cloneable. To achieve this, the
source is composed by queues which would propagate the source data.
These queues are wrapped with a reference count, which allows these
source nodes to implement Clone.