Add initial graph implementation
The bruc engine uses a topological sorted graph to compute the resulting
chart from the specification. This follows the same implementation
from Vega, where Pulse instances keep track of the state for each
of the nodes of the graph. This needs to be extended for all
the possible transforms, scale & rendering capabilities